# docs/openapi.yaml
openapi: 3.0.3
info:
  title: نظام إدارة محل السبائك والمشغولات — API
  description: |
    REST API لنظام إدارة محل الذهب والفضة: مخزون مركزي، تسعير يومي، فروع متعددة،
    نقاط بيع (POS)، حجوزات، جرد، شيفتات، تقارير، ومزامنة مع WooCommerce.

    ## المصادقة
    - جميع النقاط تحتاج `Authorization: Bearer <token>` ما عدا `/api/auth/login` و `/api/health`.
    - توكن JWT (خوارزمية HS256، صلاحية 12 ساعة) يُحصل عليه من `/api/auth/login`.
    - لكل مسؤول صلاحيات (permissions) تُحمّل مع كائن الموظف عند الدخول.
      النقاط المحمية بصلاحية محددة موثّقة بـ «Requires permission» في الوصف،
      وترجع `403 {error:"forbidden"}` عند غيابها.

    ## الأخطاء
    جميع الأخطاء تُرجع JSON على الصورة: `{"error": "<code>"}` مع اختياري
    `message` (نص عربي قابل للعرض) و `detail` (لأخطاء الاتصال) و `refs` (للحذف المقيّد).
    الكود الداخلي في `error` ثابت وقابل للبرمجة عليه.

    ## توكن الدخول
    استخدم `/api/auth/login` بالحسابات الأولية (PIN: `1234`):
    | username | الدور |
    |---|---|
    | `manager` | مدير المحل |
    | `cashier` | كاشير |
    | `social` | سوشيال |
  version: 1.0.0
  contact:
    name: Jewelry Store System
servers:
  - url: http://localhost:4001
    description: تطوير (Vite proxy)
  - url: https://your-domain.example
    description: إنتاج
...
security:
  - bearerAuth: []
paths:
...
  /api/stock-limits:
    get:
      tags: [StockLimits]
      summary: حدود المخزون (min/max) لكل فرع+معدن+عيار
      responses:
        '200':
          description: الحدود
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/StockLimit' }
    post:
      tags: [StockLimits]
      summary: إنشاء/تحديث حد مخزون (upsert بالفرع+معدن+عيار)
      description: "Requires permission: `stockcount.manage`"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [locationId, metalType]
              properties:
                locationId: { type: integer }
                metalType: { type: string, enum: [gold, silver] }
                carat: { type: string }
                minQty: { type: integer, default: 0 }
                maxQty: { type: integer }
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockLimit'
...
  /api/payment-methods:
    get:
      tags: [PaymentMethods]
      summary: طرق الدفع مع إحصاء استخدام 30 يومًا
      responses:
        '200':
          description: الطرق
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PaymentMethod' }
    post:
      tags: [PaymentMethods]
      summary: إضافة طريقة دفع
      description: "Requires permission: `settings.manage`. يُستنتج الكود من `code` أو `nameAr`. 409 `payment_methods.duplicate`."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nameAr]
              properties:
                nameAr: { type: string }
                nameEn: { type: string }
                code: { type: string }
                color: { type: string, default: '#64748b' }
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentMethod'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/payment-methods/{id}:
    patch:
      tags: [PaymentMethods]
      summary: تعديل / إيقاف طريقة دفع (لا يُحذف للحفاظ على التاريخ)
      description: "Requires permission: `settings.manage`"
      parameters:
        - { $ref: '#/components/parameters/Id' }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nameAr: { type: string }
                nameEn: { type: string }
                color: { type: string }
                isActive: { type: boolean }
                sortOrder: { type: integer }
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentMethod'
        '404':
          $ref: '#/components/responses/NotFound'

  # ------------------------------------------------------------------
  # Settings
  # ------------------------------------------------------------------
...
  /api/reports/payments:
    get:
      tags: [Reports]
      summary: المدفوعات حسب طريقة الدفع (فلاتر تاريخ/طريقة/فرع)
      parameters:
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
        - { name: method, in: query, schema: { type: string } }
        - { name: locationId, in: query, schema: { type: integer } }
      responses:
        '200':
          description: التقرير
          content:
            application/json:
              schema:
                type: object
                properties:
                  byMethod:
                    type: array
                    items:
                      type: object
                      properties:
                        code: { type: string }
                        nameAr: { type: string }
                        color: { type: string }
                        isActive: { type: boolean }
                        count: { type: integer }
                        total: { type: number }
                        collected: { type: number }
                  rows:
                    type: array
                    items: { $ref: '#/components/schemas/Invoice' }
                  summary:
                    type: object
                    properties:
                      total: { type: number }
                      count: { type: integer }
                      avgInvoice: { type: number }
  /api/reports/profitability:
    get:
      tags: [Reports]
      summary: الربحية الحقيقية (فرق سعر المعدن × الوزن + المصنعية − التكلفة)
      parameters:
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
      responses:
        '200':
          description: التقرير
          content:
            application/json:
              schema:
                type: object
                properties:
                  rows:
                    type: array
                    items:
                      type: object
                      properties:
                        invoiceNo: { type: string }
                        createdAt: { type: string, format: date-time }
                        cashierName: { type: string }
                        itemCodeSnapshot: { type: string }
                        itemNameSnapshot: { type: string }
                        metalTypeSnapshot: { type: string }
                        caratSnapshot: { type: string }
                        weightGSnapshot: { type: number }
                        metalPriceSnapshot: { type: number }
                        metalCostPrice: { type: number }
                        metalProfit: { type: number }
                        craftsmanshipCharged: { type: number }
                        cost: { type: number }
                        profit: { type: number }
                        invoiceStatus: { type: string }
                  summary:
                    type: object
                    properties:
                      totalProfit: { type: number }
                      invoiceCount: { type: integer }
  /api/reports/slow-stock:
    get:
      tags: [Reports]
      summary: المخزون الراكد (لم يُبع منذ N يوم)
      parameters:
        - { name: days, in: query, schema: { type: integer, description: افتراضيًا من الإعدادات (90) } }
      responses:
        '200':
          description: القطع الراكدة
          content:
            application/json:
              schema:
                type: object
                properties:
                  days: { type: integer }
                  rows:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/Item'
                        - type: object
                          properties:
                            locationName: { type: string }
                            categoryName: { type: string }
                            daysInStock: { type: integer }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: توكن JWT من /api/auth/login (صلاحية 12 ساعة)
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer }
      description: معرّف رقمي
  responses:
...
    NotFound:
      description: غير موجود
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: تعارض (حالة غير صالحة أو مرجع محجوب)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string, description: كود خطأ ثابت قابل للبرمجة، مثال notfound / items.in_use }
        message: { type: string, description: نص عربي قابل للعرض (اختياري) }
        detail: { type: string, description: تفاصيل إضافية (اختياري) }
        refs:
          type: object
          description: إحصاءات المراجع للحذف المقيّد (اختياري)

    # ---- Auth ----
...
    Item:
      type: object
      properties:
        id: { type: integer }
        code: { type: string }
        barcode: { type: string, nullable: true }
        name: { type: string, nullable: true }
        description: { type: string, nullable: true }
        photoUrl: { type: string, nullable: true, description: مسار /uploads/... }
        categoryId: { type: integer, nullable: true }
        categoryCode: { type: string, nullable: true }
        categoryName: { type: string, nullable: true }
        size: { type: string, nullable: true }
        metalType: { type: string, enum: [gold, silver] }
        carat: { type: string, nullable: true, description: 21 / 24 / 925 }
        weightG: { type: number }
        stoneWeightG: { type: number }
        craftsmanshipType: { type: string, enum: [fixed, percent] }
        craftsmanshipValue: { type: number }
        cost: { type: number, nullable: true }
        metalPriceAtAdd: { type: number, nullable: true }
        sourceSupplier: { type: string, nullable: true }
        status: { type: string, enum: [available, reserved, sold, in_transit] }
        physicalStatus: { type: string, enum: [new, used] }
        notes: { type: string, nullable: true }
        manufacturingVarianceG: { type: number }
        quantity: { type: integer }
        reservedQty: { type: integer }
        inTransitQty: { type: integer }
        availableQty: { type: integer, description: quantity − reserved − in_transit (مُحسّب) }
        currentLocationId: { type: integer, nullable: true }
        locationCode: { type: string, nullable: true }
        locationName: { type: string, nullable: true }
        isActive: { type: boolean }
        wcProductId: { type: integer, nullable: true }
        wcLastSyncedAt: { type: string, format: date-time, nullable: true }
        createdBy: { type: integer, nullable: true }
        createdByName: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    StockLimit:
      type: object
      properties:
        id: { type: integer }
        locationId: { type: integer }
        metalType: { type: string, enum: [gold, silver] }
        carat: { type: string, nullable: true }
        minQty: { type: integer }
        maxQty: { type: integer, nullable: true }

    # ---- Invoices ----
    Invoice:
      type: object
      properties:
        id: { type: integer }
        invoiceNo: { type: string, example: INV-20260808-0001 }
        employeeId: { type: integer }
        cashierName: { type: string }
        locationId: { type: integer }
        locationName: { type: string }
        customerId: { type: integer, nullable: true }
        customerName: { type: string, nullable: true }
        customerPhone: { type: string, nullable: true }
        metalSubtotal: { type: number }
        craftsmanshipTotal: { type: number }
        discountAmount: { type: number }
        discountReason: { type: string, nullable: true }
        discountApprovedBy: { type: integer, nullable: true }
        approvedByName: { type: string, nullable: true }
        vatPercent: { type: number }
        vatAmount: { type: number }
        total: { type: number }
        paymentMethod: { type: string }
        paymentMethodName: { type: string }
        paymentMethodColor: { type: string }
        status: { type: string, enum: [active, returned] }
        shiftId: { type: integer, nullable: true }
        isOffline: { type: boolean }
        deviceId: { type: string, nullable: true }
        returnReason: { type: string, nullable: true }
        returnedAt: { type: string, format: date-time, nullable: true }
        returnedBy: { type: integer, nullable: true }
        returnedByName: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
...
    PaymentMethod:
      type: object
      properties:
        id: { type: integer }
        code: { type: string, example: cash }
        nameAr: { type: string }
        nameEn: { type: string, nullable: true }
        color: { type: string }
        isActive: { type: boolean }
        sortOrder: { type: integer }
        invoices30d: { type: integer }
        total30d: { type: number }

    # ---- Dashboard ----
...